Add files via upload
The primary goal of this project is to create an automated system for watering plants. This system uses a soil moisture sensor to measure the moisture content of the soil and a servo motor to control the water sprinkling mechanism. The system ensures that plants receive the right amount of water based on the soil moisture level, thus preventing under or over-watering.